% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catdata}
\alias{catdata}
\title{Dummy Categorical Dataset}
\format{
A data frame with 40 rows and 8 variables:
\describe{
  \item{\code{SUBJECTNUM}}{integer: Unique subject ID}
  \item{\code{AGE}}{integer: age of participants} 
  \item{\code{SEX}}{character:sex of participants} 
  \item{\code{GROUPING}}{character: Study group/arm/treatment/IP}
  \item{\code{VISITNUMBER}}{integer: Visit number}
  \item{\code{PAIN_FREQUENCY}}{character: pain frequency category} 
  \item{\code{PAIN_FREQUENCY_SCORE}}{integer: pain frequency score} 
  \item{\code{PAIN_IMPROVED}}{integer: pain frequency improvement status} 
}
}
\usage{
catdata
}
\description{
Dummy dataset with categorical outcomes. One outcome is pain 
frequency at the end of the study (Visit 2). This is entered as frequency 
category ("PAIN_FREQUENCY") and recoded as a numeric score ("PAIN_FREQUENCY_SCORE"). 
All participants had pain severity of "Very Often" at baseline or Visit 1 
(this data is not included). Based on pain frequency at the end of the study, 
participants are classified as having improved (1) or not (0) ("PAIN_IMPROVED")
}
\keyword{datasets}
