% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkssdata}
\alias{pkssdata}
\title{Dummy PK Steady-state Dataset}
\format{
A data frame with 49 rows and 9 variables:
\describe{
  \item{\code{SUBJECTNUM}}{integer Unique subject ID}
  \item{\code{CONC}}{double Concentration of analyte (ng/mL)}
  \item{\code{TIME}}{double Time (h)}
  \item{\code{SEQ}}{character Treatment sequence ("AB" or "BA")}
  \item{\code{PERIOD}}{integer Study period (1 or 2)}
  \item{\code{TRT}}{character Treatment/Investigational Product received ("A" or "B")}
  \item{\code{GROUPING}}{character Study group. Usually the same as TRT}
  \item{\code{ANALYTE}}{character Analyte being measured}
  \item{\code{ANALYTEN}}{integer Numeric coding for analyte} 
}
}
\usage{
pkssdata
}
\description{
Dummy PK steady-state dataset for a single study participant.
}
\keyword{datasets}
