% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.desc_stats.R
\name{q.desc_stats}
\alias{q.desc_stats}
\title{Generate descriptive summaries for continuous data}
\usage{
q.desc_stats(x, groupvar="", outcomevar="", timevar="")
}
\arguments{
\item{x}{data frame}

\item{groupvar}{Variable by which to group data when generating descriptive summaries}

\item{outcomevar}{Outcome variable name. Variable itself must be at least numeric}

\item{timevar}{Time/visit variable name.}
}
\value{
Returns a tibble of results
}
\description{
Generate descriptive summaries for continuous data namely mean, standard deviation, number of observations, median, minimum, maximum
}
\examples{
id <- rep(1:5, each=4)
baseline_values <- rep(rnorm(5,15,4), each=4)
outcome <- c(rnorm(20,25,5))
visit <- as.factor(c(rep(1:4,5)))
q.data <- cbind.data.frame(id, baseline_values, outcome, visit)
q.desc_stats(q.data, groupvar="visit", outcomevar="outcome", timevar="visit")
}
\author{
Abdul Malik Sulley
}
