% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.locf.R
\name{q.locf}
\alias{q.locf}
\title{LOCF Imputation}
\usage{
q.locf(db, sliceby=c())
}
\arguments{
\item{db}{Data frame with missing values}

\item{sliceby}{Fields by which the data should be split before imputation. This ensures that each imputation is only based on the same subject's data}
}
\value{
A data frame
}
\description{
Perform "last observation carried forward" imputation on a dataset. If the first measurement is missing, the function also performs "first observation carried backward" imputation.
}
\note{
Dataset needs to be ordered by subject ID and time, and any other characteristics. This ensures that only the previous measurement is used for imputation.
}
\examples{
mybeaver <- beaver1
#order by day and time
mybeaver <- mybeaver[order(mybeaver$day, mybeaver$time),] 
#make some data missing
mybeaver$temp[c(6, 10, 22, 55, 57)] <- NA
q.locf(mybeaver, sliceby="day")
}
\seealso{
zoo::na.locf
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> April 5, 2020
}
