% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.GenDataFrame.R, R/q.n_decimals.R
\name{q.n_decimals}
\alias{q.n_decimals}
\title{Round numbers up or down}
\usage{
q.n_decimals(x, k)

q.n_decimals(x, k)
}
\arguments{
\item{x}{Number}

\item{k}{Number of decimal places}
}
\value{
Returns a string

Returns a string
}
\description{
Round numbers up or down to specified decimal places and return a string

Round numbers up or down to specified decimal places and return a string
}
\examples{
q.n_decimals(2.239849223, 3)
## 2.240
q.n_decimals(2.239849223, 3)
## 2.240
}
\author{
Abdul Malik Sulley

Abdul Malik Sulley
}
