% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.netconcs.R
\name{q.netconcs}
\alias{q.netconcs}
\title{Generate net analyte concentrations}
\usage{
q.netconcs(x, suffix="")
}
\arguments{
\item{x}{Data frame. See example for format}

\item{suffix}{string/text to be included in the list item name}
}
\value{
Returns a data frame.
}
\description{
Generate net analyte concentrations
}
\details{
Generate net analyte concentrations.
}
\examples{
x <- pkdata
q.netconcs(x)

}
\author{
Abdul Malik Sulley <asulley@uwo.ca> Feb 12, 2021
}
