% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.newPayload.R
\name{q.newPayload}
\alias{q.newPayload}
\title{Internal function newPayload}
\usage{
q.newPayload()
}
\value{
A list called q.payload
}
\description{
Create a list() object to hold data frames and table titles from other functions
}
\examples{
q.newPayload()
}
