% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.pooled_sd.R
\name{q.pooled_sd}
\alias{q.pooled_sd}
\title{Pooled Standard Deviation}
\usage{
q.pooled_sd(sds=c(), ns=c())
}
\arguments{
\item{sds}{A vector containing standard deviations to be pooled}

\item{ns}{A vector of length \code{length(sds)}, containing the sample sizes matching each standard deviation in \code{sds}}
}
\value{
Pooled standard deviation
}
\description{
Calculate the pooled standard deviation given a set of standard deviations and sample sizes
}
\examples{
sds <- c(21.1, 32.2, 43.3)
ns <- c(40, 37, 39)
q.pooled_sd(sds, ns)
}
