% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.recodeNA.R
\name{q.recodeNA}
\alias{q.recodeNA}
\title{Recode specified entries as NA (missing)}
\usage{
q.recodeNA(x, pattern=c(-77, -99, -55, -33))
}
\arguments{
\item{x}{A data frame}

\item{pattern}{A vector of numbers or character strings to be replaced with NA. Default: -77, -99, -55, -33}
}
\value{
Returns a data frame
}
\description{
Recode specified values as missing (NA)
}
\examples{
id <- c(1, 2, 3, 4, 5)
value <- c(25.2, 33.3, -99, 13.0, -77)
dframe <- data.frame(id, value)
q.recodeNA(x=dframe, pattern=c(-77, -99))
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> May 11, 2020
}
