% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.rename_headers.R
\name{q.rename_headers}
\alias{q.rename_headers}
\title{Internal function: Format table headers}
\usage{
q.rename_headers(x, type="default")
}
\arguments{
\item{x}{data frame}

\item{type}{Type of analysis. Specify one of these options: \code{c("longitudinal", "categorical", "default")}}
}
\value{
Returns a tibble of results
}
\description{
Format table headers depending on the type of analysis being conducted.
}
\details{
For categorical variables, adds "n (%)" to headers. For Continuous variables, adds the following "Mean+-SD (n)", "Median (Min - Max)", "Within Group P-value"
}
\examples{
df.header <- c("#", "Group A", "Group B", "p.val")
row1 <- c("Day 7", 20, 50, 0.040)
row2 <- c("Day 14", 3, 12, 0.010)
df <- rbind.data.frame(row1, row2)
names(df) <- df.header
q.rename_headers(df)
}
\author{
Abdul Malik Sulley
}
