% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.samp.size.p.R
\name{q.samp.size.p}
\alias{q.samp.size.p}
\title{Sample size for comparing proportions}
\usage{
q.samp.size.p(p1=NULL, p2=NULL, dropout=0.2, arms=2, 
               sig.level = 0.05 , power = 0.8, alternative = "two.sided")
}
\arguments{
\item{p1}{Proportion for the product/intervention of interest}

\item{p2}{Proportion for comparator/placebo}

\item{dropout}{Dropout rate to be accounted for in sample size calculation. Default is 0.2}

\item{arms}{Number of armes in the planned study}

\item{sig.level}{Significance level. Default is 0.05}

\item{power}{Power. Default is 0.8}

\item{alternative}{A charater string specifying the alternative hypothesis. Options:  "two.sided" (default), "greater" or "less" (see pwr.2p.test)}
}
\value{
A list containing the ratio of proportions, number of completers per arm, total sample size per arm, total sample size for the study and number of study arms
}
\description{
Calculate sample size for comparing proportions based on \code{pwr::pwr.2p.test}
}
\examples{
p1 <- 0.8
p2 <- 0.5
q.samp.size.p(p1, p2, dropout=0.2, arms=2, 
              sig.level = 0.05 , power = 0.8, alternative = "two.sided" )
}
\seealso{
pwr.2p.test
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> May 8, 2020
}
