% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.samp.size.t.R
\name{q.samp.size.t}
\alias{q.samp.size.t}
\title{Sample size for comparing means}
\usage{
q.samp.size.t(refmean=NULL, compmean=NULL, SD=NULL, dropout=0.2, arms=2, 
             sig.level = 0.05 , power = 0.8, type = "two.sample", 
             alternative ="two.sided")
}
\arguments{
\item{refmean}{Mean for the product/intervention of interest}

\item{compmean}{Mean for comparator/placebo}

\item{SD}{Pooled standard deviation for calculation of Cohen's d}

\item{dropout}{Dropout rate to be accounted for in sample size calculation. Default is 0.2}

\item{arms}{Number of armes in the planned study}

\item{sig.level}{Significance level. Default is 0.05}

\item{power}{Power. Default is 0.8}

\item{type}{A character string specifying the type of test. Options: "one.sample", "two.sample" (default) or "paired" (see pwr.t.test)}

\item{alternative}{A charater string specifying the alternative hypothesis. Options:  "two.sided" (default), "greater" or "less" (see pwr.t.test)}
}
\value{
A list containing the mean difference, mean ratio, number of completers per arm, total sample size per arm, total sample size for the study and number of study arms
}
\description{
Calculate sample size for a difference in mean based on pwr.t.test from the pwr package
}
\examples{
sds <- c(21.1, 32.2, 43.3)
ns <- c(40, 37, 39)
SD <- q.pooled_sd(sds, ns)
refmean <- 57.2
compmean <- 22.8
q.samp.size.t(refmean, compmean, SD, dropout=0.2, 
type = "two.sample", alternative ="two.sided", arms=2 )
}
\seealso{
pwr.t.test
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> 22 April, 2020
}
