% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.substring_right.R
\name{q.substring_right}
\alias{q.substring_right}
\title{Substring of a Character Vector from Right End of the String}
\usage{
q.substring_right(x, n)
}
\arguments{
\item{x}{string}

\item{n}{number of characters from the right end of the string to extract}
}
\value{
A character vector of length n
}
\description{
Extract a substring n characters from the right end of a string
}
\details{
Extract a substring n characters from the right end of a string. This is useful when traversing a vector, list or data frame with strings of different lengths.
}
\examples{
q.substring_right("twelve", 3)
# "lve"
}
\author{
Abdul Malik Sulley
}
