\name{qwickr-package}
\alias{qwickr-package}
\alias{qwickr}
\docType{package}
\title{
\packageTitle{qwickr}
}
\description{
\packageDescription{qwickr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qwickr}
\packageIndices{qwickr}

}
\author{
\packageAuthor{qwickr}
Maintainer: \packageMaintainer{qwickr}
}
\references{
To be updated
}

\keyword{ package }
\keyword{ qwickr-package }
\keyword{ qwickr }

