% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwickr.cat.R
\name{qwickr.cat}
\alias{qwickr.cat}
\title{Summarize for categorical data}
\usage{
qwickr.cat(x, outcomevar="", groupvar="", timevar="", specgroups=NULL, 
          filesuffix="", t.title="", genpvalues=T, runpairwise=F, 
          exportfile=c(), exportpath="")
}
\arguments{
\item{x}{Data frame}

\item{outcomevar}{Name of outcome variable \code{(string)}}

\item{groupvar}{Name of the group variable \code{(string)}}

\item{timevar}{Name of the time variable \code{(string)}}

\item{specgroups}{Do you want to compare a subset of groups within your dataset? Specify the group names here. For example, \code{specgroups=c("Group A", "Group B", "Group C")}. If NULL, all groups will be included in the analysis.}

\item{filesuffix}{A string to be appended to the name of the returned .csv or .doc file}

\item{t.title}{Table title. Exported together with a dataframe of the output as a list. This list can then be exported to Word using the \code{q.write.to.word()} command}

\item{genpvalues}{Should a Chi Square or Fisher's Exact test be performed? \code{T/F}. If true, the function will determine which test to perform}

\item{runpairwise}{Run pairwise comparisons for each unique group pair? \code{T/F}}

\item{exportfile}{Export the output to file? Options: \code{c(".csv", ".doc")}. See \code{q.write.to.word, stats::write.csv }}

\item{exportpath}{Path relative to the working directory where exported files will be saved e.g. "OUTPUT" Do not begin or end with a backslash. If left empty, file will be exported to the working directory.}
}
\value{
Returns a data frame of counts and percentages for each study arm and an associated p-value for each study time point in a repeated measures design.
}
\description{
Run descriptive summaries for categorical data with the option to include hypothesis testing using the Chi Square test or Fisher's Exact test as appropriate.
}
\details{
The function will automatically determine whether to use Chi Square or Fisher's Exact test based on the nature of the data.
}
\examples{
group <- rep(c("A", "B"), 10)
gender <- rep(c(1,1,0,0), 5)
time <- rep(1, 10)
df <- data.frame(group, gender, time)
qwickr.cat(x=df, outcomevar="gender", groupvar = "group", timevar = "time")
}
\seealso{
stats::fisher.test(), q.write.to.word, utils::write.csv
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> May 7, 2020
}
