% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwickr.pkbe.R
\name{qwickr.pkbe}
\alias{qwickr.pkbe}
\title{Calculate Bioequivalence with 90\% CI}
\usage{
qwickr.pkbe(x, outcomevars=c(), design="crossover")
}
\arguments{
\item{x}{Data frame. See example for format}

\item{outcomevars}{vector of variable names/column titles based on which BE will be calculated e.g.\code{c("LOGAUC", "LOGCMAX")}}

\item{design}{specify the study design. Options: c("parallel", "crossover").}
}
\value{
Returns a list of data frames - one data frame per analyte.
}
\description{
Calculate bioequivalence along with 90\% confidence intervals and coefficient of variation
}
\details{
Calculate bioequivalence along with 90\% confidence intervals and coefficient of variation
}
\examples{
x <- qwickr.pkparams(pkdata)
pkbe.output <- qwickr.pkbe(x[[1]], outcomevars = c("LOGAUC", "LOGCMAX"), design = "crossover")
print(pkbe.output)
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> Feb 18, 2021
}
