% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwickr.pkconcs.R
\name{qwickr.pkconcs}
\alias{qwickr.pkconcs}
\title{Generate PK concentration listings}
\usage{
qwickr.pkconcs(db, suffix="")
}
\arguments{
\item{db}{Data frame. See example for format}

\item{suffix}{string/text to be included in the list item name}
}
\value{
Returns a list of data frames.
}
\description{
Generate PK concentration listings
}
\details{
Generates PK concentration listings according to analyte and study group. It generates tables for both raw concentrations and change from baseline/pre-dose.
}
\examples{
db <- pkdata

qwickr.pkconcs(db)
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> Feb 8, 2021
}
