% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwickr.pkfe.R
\name{qwickr.pkfe}
\alias{qwickr.pkfe}
\title{Generate Fixed Effects Tables for PK Parameters}
\usage{
qwickr.pkfe(x, outcomevar, groupvar,  design="", assume.normal.dist=F)
}
\arguments{
\item{x}{Data frame}

\item{outcomevar}{Name of outcome variable \code{(string)}}

\item{groupvar}{Name of the group variable \code{(string)}}

\item{design}{specify the study design. Options: c("parallel", "crossover"). Default: "crossover"}

\item{assume.normal.dist}{assume that the data is normally distributed? (T/F) Default: `FALSE`, the Shapiro Wilk test will be used to assess normality of the data. Data will be ranked if not normally distributed.}
}
\value{
Returns a data frame of counts and percentages for each study arm and an associated p-value for each study time point in a repeated measures design.
}
\description{
Run descriptive summaries and generate fixed effects tables
}
\details{
Uses linear mixed effects model to compare PK parameters between groups and generate fixed effects tables as per [Health Canada](https://www.canada.ca/content/dam/hc-sc/documents/services/drugs-health-products/drug-products/applications-submissions/guidance-documents/bioavailability-bioequivalence/conduct-analysis-comparative.pdf) format.
}
\examples{
x <- qwickr.pkparams(pkdata)
pkfe.output <- qwickr.pkfe(x[[1]], outcomevar = "AUC", groupvar = "GROUPING")
}
\seealso{
stats::fisher.test(), q.write.to.word, utils::write.csv
}
\author{
Abdul Malik Sulley <asulley@uwo.ca> May 7, 2020
}
