% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwickr.pkparams.R
\name{qwickr.pkparams}
\alias{qwickr.pkparams}
\title{Generate PK parameters for each subject}
\usage{
qwickr.pkparams(db, suffix="")
}
\arguments{
\item{db}{Data frame. See example for format}

\item{suffix}{string/text to be included in the list item name}
}
\value{
Returns a list containing data frames of PK parameters for each analyte and study group, and one data frame containing data for all participants.
}
\description{
Generate PK parameters for each subject
}
\details{
Generates PK parameters for each subject, analyte and study group.
}
\examples{
db <- pkdata
qwickr.pkparams(db)

}
\author{
Abdul Malik Sulley <asulley@uwo.ca> Feb 12, 2021
}
