% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmdata}
\alias{rmdata}
\title{Dummy Repeated Measures Dataset}
\format{
A data frame with 24 rows and 4 variables:
\describe{
  \item{\code{SUBJECTNUM}}{integer: Unique subject ID}
  \item{\code{GROUPING}}{character: Study group/arm/treatment/IP}
  \item{\code{VISITNUMBER}}{integer: Visit number}
  \item{\code{BIOMARKER}}{double: Biomarker used to assess study endpoints} 
}
}
\usage{
rmdata
}
\description{
Dummy repeated measures dataset.
}
\keyword{datasets}
